Added naive extension of tubes in TubeInit

+++ CompuCell3D/core/CompuCell3D/steppables/TubeFieldInitializer/TubeFieldInitializer.h
@@ -82,9 +82,9 @@ namespace CompuCell3D {
 
         std::vector<double> crossProductVec(const std::vector<double>& v1, const std::vector<double>& v2);
 
-        float magnitude(Point3D p);
+        double magnitude(Point3D p);
 
-        float distanceToLine(Point3D line_point1, Point3D line_point2, Point3D point);
+        double distanceToLine(Point3D line_point1, Point3D line_point2, Point3D point);
